feat: mods in pre-U18.18 builds + equipment features in pre-U24.4 builds (#3040)
This implements support for mods in pre-U18.18  (The Silver Grove) builds, as well as support for Orokin Catalysts/Reactors, Forma, Exilus Adapters, and gilding modular weapons in pre-U24.4 builds. And should also fix #3036.

- Pre-Endo fusion system (consuming Fusion Cores and other mods to rank up the target), handles ranked Fusion Core mission drops.
- Attaching/detaching mods in U7-U8.
- Transmutation works but the game crashes if it rolls a mod that doesn't exist in that build.
- Riven mods work in U19 builds when added through the WebUI, but I can't get veiled Rivens to appear in the inventory in old builds yet (they probably changed the item paths when they made them stackable) to test if activating/unveiling them works, but rerolling them works.

This is a decently large PR that should be tested for regressions before merging, I have done a lot of thorough testing myself as I've worked on it, but I may have missed an edge-case somewhere.
And if there is anything I can improve code-wise in all these changes please reach out.

Co-authored-by: Sainan <63328889+Sainan@users.noreply.github.com>
Reviewed-on: https://onlyg.it/OpenWF/SpaceNinjaServer/pulls/3040
Reviewed-by: Sainan <63328889+sainan@users.noreply.github.com>
Co-authored-by: VoltPrime <subsonicjackal@gmail.com>
Co-committed-by: VoltPrime <subsonicjackal@gmail.com>